requirements.txt
selenium
pytest
allure-pytest
allure-python-commons
